Replaced bzip2 with gzip and fixed instructions
I had left some inconsistencies in using gzip vs. bzip2, and I had not
added the `errlog` file to the list of files that should be included in
the submission repository. That's been fixed in the readme, and I have
made the submission files in the NJIT repo more consistent with the
instructions.