Updated Results and Evals using Pt2 Assessments
TREC-IS annotation for 2021a has been split into stages to support
additional annotation. Now that the second stage has been finished,
we're releasing the stage-1 labels and have updated the leaderboard with
the stage-2 evaluations (which contain more assessments than the
stage-1).

Results are largely consistent with stage-1 assessment: the `ens`
submission from `ucd-cs` remains the top scorer for all metrics except
for actionable F1 for info-types and accuracy.